- CodeSnippets 1.1.74 2007/04/14 - Remove Gentoo linux AMD64 ambiguity on finddlg wxComboBox call
git-svn-id: https://svn.code.sf.net/p/codeblocks/code/trunk@3855 2a5c6006-c6dd-42ca-98ab-0921f2732cef

+++ src/plugins/contrib/codesnippets/finddialogs.cpp
@@ -117,7 +117,8 @@ myFindReplaceDlg::myFindReplaceDlg (wxWindow *parent,
     wxString finddir;
     m_finddir = new wxComboBox (this, -1, finddir,
                                 wxDefaultPosition, wxSize(200, -1),
-                                0, NULL);
+                                //0, NULL); //AMD64 ambiguity betw int vs wxArray
+                                (int)0, NULL);
     m_fdirsizer->Add (m_finddir, 1, wxALIGN_CENTRE_VERTICAL);
     m_findpane->Show (m_fdirsizer, false);
     m_findpane->Add (m_fdirsizer, 0, wxEXPAND | wxBOTTOM, 6);
